+++ ctview/map_class/post_processing.py
@@ -1,5 +1,7 @@
 import os
 
+import numpy as np
+from osgeo import gdal, gdal_array
 from osgeo_utils import gdal_fillnodata
 
 from ctview import utils_gdal
@@ -54,3 +56,42 @@ def add_color_to_raster(in_raster: str, output_file: str, LUT: str):
 
     if not os.path.exists(output_file):  # if raster not create, next step with fail
         raise FileNotFoundError(f"{output_file} not found")
+
+
+def smooth_class_array(class_map_array: np.array, nconnectedness: int, threshold: int):
+    """This method uses Gdal Sieve in order to remove defects in the class map
+
+    Args:
+        class_map_array (np.array): raster array to smooth
+        nconnectednass (int): gdal sieve parameter 4 indicating that diagonal
+            pixels are not considered directly adjacent for
+            polygon membership purposes or 8 indicating they are
+        threshold (int): raster polygons with sizes smaller than this will be merged into their largest neighbour
+    Returns:
+        class_map_array (np.array): raster array smoothed
+    """
+    rows, cols = class_map_array.shape
+    driver = gdal.GetDriverByName("MEM")
+    dataset = driver.Create("", cols, rows, 1, gdal.GDT_Byte)
+    band = dataset.GetRasterBand(1)
+    gdal_array.BandWriteArray(band, class_map_array)
+    gdal.SieveFilter(band, None, band, threshold, nconnectedness)
+    class_map_array = band.ReadAsArray()
+
+    return class_map_array
+
+
+def post_processing(class_map_array: np.array, nconnectedness: int, threshold: int):
+    """This method groups all post process operated on the class map
+        - Smoothing
+        - Merge with condition (TODO)
+
+    Args:
+        class_map_array (np.array): raster array to post process
+        nconnectednass (int): smoothing parameter (option for diagonal pixels)
+        threshold (int): smoothing parameter (size of minimum pixel size of polygon)
+    Returns:
+        class_map_array_post_processed (np.array): output raster array
+    """
+    class_map_array_post_processed = smooth_class_array(class_map_array, nconnectedness, threshold)
+    return class_map_array_post_processed