+++ test/map_class/test_post_processing.py
@@ -2,10 +2,16 @@ import os
 import shutil
 from pathlib import Path
 
+import numpy as np
+import pytest
 import rasterio
 from osgeo import gdal
 
-from ctview.map_class.post_processing import add_color_to_raster, fill_gaps_raster
+from ctview.map_class.post_processing import (
+    add_color_to_raster,
+    fill_gaps_raster,
+    smooth_class_array,
+)
 
 gdal.UseExceptions()
 
@@ -65,3 +71,79 @@ def test_add_color_to_raster():
         assert band_R[17, 14] == 64  # artefact (classe 65) -> color defined in LUT
         assert band_G[17, 14] == 0  # artefact (classe 65) -> color defined in LUT
         assert band_B[17, 14] == 128  # artefact (classe 65) -> color defined in LUT
+
+
+@pytest.mark.parametrize(
+    """class_map_array, nconnectedness, thresehold, expected_result""",
+    [
+        (
+            np.array(
+                [
+                    [6, 6, 2, 2],
+                    [6, 6, 2, 2],
+                    [2, 2, 6, 2],
+                    [2, 2, 2, 2],
+                ]
+            ),
+            1,
+            1,
+            (
+                np.array(
+                    [
+                        [6, 6, 2, 2],
+                        [6, 6, 2, 2],
+                        [2, 2, 6, 2],
+                        [2, 2, 2, 2],  # No smoothing with these parameters
+                    ]
+                )
+            ),
+        ),
+        (
+            np.array(
+                [
+                    [6, 6, 2, 2],
+                    [6, 6, 2, 2],
+                    [2, 2, 6, 2],
+                    [2, 2, 2, 2],
+                ]
+            ),
+            4,
+            4,
+            (
+                np.array(
+                    [
+                        [6, 6, 2, 2],
+                        [6, 6, 2, 2],
+                        [2, 2, 2, 2],
+                        [2, 2, 2, 2],  # Smoothing the only pixel 6 in pixels 2
+                    ]
+                )
+            ),
+        ),
+        (
+            np.array(
+                [
+                    [6, 6, 2, 2],
+                    [6, 6, 2, 2],
+                    [2, 2, 6, 2],
+                    [2, 2, 2, 2],
+                ]
+            ),
+            8,
+            4,
+            (
+                np.array(
+                    [
+                        [6, 6, 2, 2],
+                        [6, 6, 2, 2],
+                        [2, 2, 6, 2],
+                        [2, 2, 2, 2],  # No smoothing because nconnectedness = 8 accepts diagonal pixels into polygons
+                    ]
+                )
+            ),
+        ),
+    ],
+)
+def test_smooth_class_array(class_map_array, nconnectedness, thresehold, expected_result):
+    class_map_array_smooth = smooth_class_array(class_map_array, nconnectedness, thresehold)
+    assert np.array_equal(class_map_array_smooth, expected_result)